OrderIntentConveyor factory method on package object

+++ src/main/scala/io/liquirium/core/OperationRequestId.scala
@@ -2,7 +2,7 @@ package io.liquirium.core
 
 sealed trait OperationRequestId
 
-case class CompoundOperationRequestId(botId: BotId, requestIndex: Long) extends OperationRequestId {
+case class CompoundOperationRequestId(botId: BotId, requestIndex: Int) extends OperationRequestId {
 
   override def toString: String = botId.toString + "-" + requestIndex.toString
